requirements.txt
Django==2.1.7
pytz==2018.9
django-betterforms==1.2